Task 1.1-4 & 2, Validation, Middleware, Filtres, logger, Unit tests added + small fixes

+++ GameStore.WebApi/GameStore.WebApi/Controllers/CommentsController.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -19,36 +20,58 @@ namespace GameStore.WebApi.Controllers
         {
             _commentService = commentService;
         }
-        
+
         [HttpGet]
         [Route("{gameId}")]
-        public async Task<IEnumerable<BLComment>> Get(int gameId)
+        public async Task<IActionResult> Get(int gameId)
         {
-            return await _commentService.GetCommentsForPostAsync(gameId);
+            IEnumerable<BLComment> comments = await _commentService.GetCommentsForPostAsync(gameId);
+            if (comments == null || !comments.Any())
+                return NotFound();
+            else
+                return Ok(comments);
         }
 
         [HttpPost]
         [Route("{gameId}")]
-        public async Task Post(int gameId, [FromBody]BLComment comment)
+        public async Task<IActionResult> Post(int gameId, [FromBody]BLComment comment)
         {
+            if (gameId <= 0 || comment == null || comment.Body.Length == 0 || comment.GameId < 0 ||
+                comment.Name.Length == 0)
+                return BadRequest("Wrong arguments for model creation");
+
             comment.GameId = gameId;
             await _commentService.AddAsync(comment);
+            return Created("/api/comments/"+gameId, comment);
         }
 
         [HttpPost]
         [Route("{gameId}/{commentId}")]
-        public async Task AnswerToComment(int gameId, int commentId, [FromBody]BLComment comment)
+        public async Task<IActionResult> AnswerToComment(int gameId, int commentId, [FromBody]BLComment comment)
         {
+            if (gameId <= 0 || comment == null || comment.Body.Length == 0 || comment.GameId < 0 ||
+                comment.Name.Length == 0 || commentId <= 0)
+                return BadRequest("Wrong arguments for model creation");
+
             comment.GameId = gameId;
+            if ((await _commentService.GetById(commentId))?.Body == null)
+                return NotFound();
+
             comment.ParentCommentId = commentId;
             await _commentService.AddAsync(comment);
+
+            return Created("api/comments/"+gameId+"/"+commentId, comment);
         }
 
         [HttpDelete]
         [Route("{id}")]
-        public async Task Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
+            if (id <= 0 || await _commentService.GetById(id) == null)
+                return NotFound();
+
             await _commentService.DeleteAsync(id);
+            return Ok();
         }
 
     }
